third_party/WebKit/LayoutTests/fast/scroll-behavior/main-frame-interrupted-scroll-expected.txt

Test that interrupting a smooth scroll on the main frame works with both scroll behaviors and with input

This is a testharness.js-based test.
FAIL instant scroll assert_equals: expected 3500 but got 7000
FAIL smooth scroll assert_equals: expected 3500 but got 7000
FAIL touch scroll assert_equals: expected 3500 but got 7000
FAIL wheel scroll assert_equals: expected 3500 but got 7000
Harness: the test ran to completion.
